skriva model/pytorch_model.bin berre éin gong

+++ setup.sh
@@ -1,5 +1,10 @@
 #!/bin/sh
-cat model/x* > model/pytorch_model.bin
-rm model/x*
-pip3 install -r requirements.txt
 
+# Skapa model/pytorch_model.bin berre éin gong, elles vert det overskrive.
+ls model/x* &>/dev/null
+if [[ "$?" -eq 0 ]]; then
+    cat model/x* > model/pytorch_model.bin
+    rm model/x*
+fi
+
+pip3 install -r requirements.txt
